cambios desde sion

+++ ServicesCI/Dtos/cardsDto/cardsCreateDto.cs
@@ -2,7 +2,7 @@ namespace ControlIDMvc.ServicesCI.Dtos.cardsDto
 {
     public class cardsCreateDto
     {
-        public int value { get; set; }
+        public long value { get; set; }
         public int user_id { get; set; }
 
     }
